+++ packages/components/src/phoneNumberInput/utils/isStringNumeric/index.js
@@ -1,0 +1,1 @@
+export { isStringNumeric } from './isStringNumeric';