+++ packages/components/src/phoneNumberInput/utils/isStringNumeric/isStringNumeric.js
@@ -1,0 +1,1 @@
+export const isStringNumeric = value => /^\+?[\d-\s]+$/.test(value);